+++ java-src/preflex/instrument/concurrent/ExecutorServiceWrapper.java
@@ -20,47 +20,46 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import preflex.instrument.EventHandlerFactory;
 import preflex.instrument.task.CallTask;
 import preflex.instrument.task.CallTask1;
 import preflex.instrument.task.CallTask2;
 import preflex.instrument.task.CallTask3;
-import preflex.instrument.task.InstrumentingWrapper;
 import preflex.instrument.task.RunTask;
+import preflex.instrument.task.Wrapper;
 
 public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent> implements ExecutorService {
 
     private final ExecutorService orig;
 
     private final ConcurrentEventFactory<ThreadPoolEvent, FutureEvent, ExecutionEvent> eventFactory;
-    private final InstrumentingWrapper<ThreadPoolEvent> callableSubmitWrapper;
-    private final InstrumentingWrapper<ThreadPoolEvent> multipleSubmitWrapper;
-    private final InstrumentingWrapper<ThreadPoolEvent> runnableSubmitWrapper;
-    private final InstrumentingWrapper<ThreadPoolEvent> shutdownRequestWrapper;
+    private final Wrapper<ThreadPoolEvent> callableSubmitWrapper;
+    private final Wrapper<ThreadPoolEvent> multipleSubmitWrapper;
+    private final Wrapper<ThreadPoolEvent> runnableSubmitWrapper;
+    private final Wrapper<ThreadPoolEvent> shutdownRequestWrapper;
 
-    private final EventHandlerFactory<ExecutionEvent> callableExecutionEventHandlerFactory;
-    private final EventHandlerFactory<ExecutionEvent> runnableExecutionEventHandlerFactory;
-    private final EventHandlerFactory<FutureEvent>    futureCancelEventHandlerFactory;
-    private final EventHandlerFactory<FutureEvent>    futureResultEventHandlerFactory;
+    private final Wrapper<ExecutionEvent> callableExecutionWrapper;
+    private final Wrapper<ExecutionEvent> runnableExecutionWrapper;
+    private final Wrapper<FutureEvent>    futureCancelWrapper;
+    private final Wrapper<FutureEvent>    futureResultWrapper;
 
     private final CallableDecorator<?>    callableDecorator;
     private final RunnableDecorator       runnableDecorator;
 
     public ExecutorServiceWrapper(ExecutorService threadPool,
             ConcurrentEventFactory<ThreadPoolEvent, FutureEvent, ExecutionEvent> eventFactory,
-            ConcurrentEventHandlerFactory<ThreadPoolEvent, FutureEvent, ExecutionEvent> eventHandlerFactories,
+            ConcurrentTaskWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent> concurrentTaskWrappers,
             CallableDecorator<?> callableDecorator,
             RunnableDecorator runnableDecorator) {
         this.orig = threadPool;
         this.eventFactory = eventFactory;
-        this.callableSubmitWrapper = new InstrumentingWrapper<>(eventHandlerFactories.callableSubmitHandlerFactory);
-        this.multipleSubmitWrapper = new InstrumentingWrapper<>(eventHandlerFactories.multipleSubmitHandlerFactory);
-        this.runnableSubmitWrapper = new InstrumentingWrapper<>(eventHandlerFactories.runnableSubmitHandlerFactory);
-        this.shutdownRequestWrapper = new InstrumentingWrapper<>(eventHandlerFactories.shutdownRequestHandlerFactory);
-        this.callableExecutionEventHandlerFactory = eventHandlerFactories.callableExecuteHandlerFactory;
-        this.runnableExecutionEventHandlerFactory = eventHandlerFactories.runnableExecuteHandlerFactory;
-        this.futureCancelEventHandlerFactory = eventHandlerFactories.futureCancelHandlerFactory;
-        this.futureResultEventHandlerFactory = eventHandlerFactories.futureResultHandlerFactory;
+        this.callableSubmitWrapper = concurrentTaskWrappers.callableSubmitWrapper;
+        this.multipleSubmitWrapper = concurrentTaskWrappers.multipleSubmitWrapper;
+        this.runnableSubmitWrapper = concurrentTaskWrappers.runnableSubmitWrapper;
+        this.shutdownRequestWrapper = concurrentTaskWrappers.shutdownRequestWrapper;
+        this.callableExecutionWrapper = concurrentTaskWrappers.callableExecuteWrapper;
+        this.runnableExecutionWrapper = concurrentTaskWrappers.runnableExecuteWrapper;
+        this.futureCancelWrapper = concurrentTaskWrappers.futureCancelWrapper;
+        this.futureResultWrapper = concurrentTaskWrappers.futureResultWrapper;
         this.callableDecorator = callableDecorator;
         this.runnableDecorator = runnableDecorator;
     }
@@ -69,7 +68,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
     public void execute(final Runnable runnable) {
         final Runnable decoratedRunnable = runnableDecorator.wrapRunnable(runnable);
         final RunnableWrapper<?> instrumentedTask = new RunnableWrapper<>(
-                decoratedRunnable, eventFactory, runnableExecutionEventHandlerFactory);
+                decoratedRunnable, eventFactory, runnableExecutionWrapper);
         final ThreadPoolEvent event = eventFactory.runnableSubmissionEvent(decoratedRunnable);
         final RunTask task = new RunTask() {
             @Override
@@ -125,14 +124,14 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
         final CallableDecorator<T> typedCallableDecorator = (CallableDecorator<T>) callableDecorator;
         final SharedContextCallable<?, T> decoratedTask = typedCallableDecorator.wrapCallable(callable);
         final CallableWrapper<T, ?> instrumentedTask = new CallableWrapper<>(decoratedTask,
-                eventFactory, callableExecutionEventHandlerFactory);
+                eventFactory, callableExecutionWrapper);
         final ThreadPoolEvent event = eventFactory.callableSubmissionEvent(decoratedTask);
         final CallTask<Future<T>> task = new CallTask<Future<T>>() {
             @Override
             public Future<T> call() {
                 final Future<T> future = orig.submit(instrumentedTask);
                 return new FutureWrapper<>(decoratedTask.wrapFuture(future),
-                        eventFactory, futureCancelEventHandlerFactory, futureResultEventHandlerFactory);
+                        eventFactory, futureCancelWrapper, futureResultWrapper);
             }
         };
         return callableSubmitWrapper.call(event, task);
@@ -142,7 +141,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
     public <T> Future<T> submit(final Runnable runnable, final T result) {
         final SharedContextRunnable<?> decoratedRunnable = runnableDecorator.wrapRunnable(runnable);
         final RunnableWrapper<?> instrumentedTask = new RunnableWrapper<>(
-                decoratedRunnable, eventFactory, runnableExecutionEventHandlerFactory);
+                decoratedRunnable, eventFactory, runnableExecutionWrapper);
         final ThreadPoolEvent event = eventFactory.runnableSubmissionEvent(decoratedRunnable);
         final CallTask<Future<T>> task = new CallTask<Future<T>>() {
             @Override
@@ -151,7 +150,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
                 @SuppressWarnings("unchecked")
                 final Future<T> decoratedFuture = (Future<T>) decoratedRunnable.wrap(future);
                 return new FutureWrapper<>(decoratedFuture,
-                        eventFactory, futureCancelEventHandlerFactory, futureResultEventHandlerFactory);
+                        eventFactory, futureCancelWrapper, futureResultWrapper);
             }
         };
         return runnableSubmitWrapper.call(event, task);
@@ -161,7 +160,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
     public Future<?> submit(final Runnable runnable) {
         final SharedContextRunnable<?> decoratedRunnable = runnableDecorator.wrapRunnable(runnable);
         final RunnableWrapper<?> instrumentedTask = new RunnableWrapper<>(
-                decoratedRunnable, eventFactory, runnableExecutionEventHandlerFactory);
+                decoratedRunnable, eventFactory, runnableExecutionWrapper);
         final ThreadPoolEvent event = eventFactory.runnableSubmissionEvent(decoratedRunnable);
         final CallTask<Future<?>> task = new CallTask<Future<?>>() {
             @Override
@@ -170,7 +169,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
                 @SuppressWarnings("unchecked")
                 final Future<Object> futureObject = (Future<Object>) future;
                 return new FutureWrapper<>(decoratedRunnable.wrap(futureObject),
-                        eventFactory, futureCancelEventHandlerFactory, futureResultEventHandlerFactory);
+                        eventFactory, futureCancelWrapper, futureResultWrapper);
             }
         };
         return runnableSubmitWrapper.call(event, task);
@@ -186,7 +185,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
         for (Callable<T> each: tasks) {
             final SharedContextCallable<?, T> decoratedTask = typedCallableDecorator.wrapCallable(each);
             final CallableWrapper<T, ExecutionEvent> eachInstrumentedTask = new CallableWrapper<>(
-                    decoratedTask, eventFactory, callableExecutionEventHandlerFactory);
+                    decoratedTask, eventFactory, callableExecutionWrapper);
             decoratedTasks.add(decoratedTask);
             instrumentedTasks.add(eachInstrumentedTask);
         }
@@ -201,7 +200,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
                 for (Future<T> each: futures) {
                     final Future<T> instrumentedFuture = new FutureWrapper<>(
                             decoratedTasks.get(i).wrapFuture(each), eventFactory,
-                            futureCancelEventHandlerFactory, futureResultEventHandlerFactory);
+                            futureCancelWrapper, futureResultWrapper);
                     futWrappers.add(instrumentedFuture);
                     i++;
                 }
@@ -222,7 +221,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
         for (Callable<T> each: tasks) {
             final SharedContextCallable<?, T> decoratedTask = typedCallableDecorator.wrapCallable(each);
             final CallableWrapper<T, ExecutionEvent> eachInstrumentedTask = new CallableWrapper<>(
-                    decoratedTask, eventFactory, callableExecutionEventHandlerFactory);
+                    decoratedTask, eventFactory, callableExecutionWrapper);
             decoratedTasks.add(decoratedTask);
             instrumentedTasks.add(eachInstrumentedTask);
         }
@@ -237,7 +236,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
                 for (Future<T> each: futures) {
                     final Future<T> instrumentedFuture = new FutureWrapper<>(
                             decoratedTasks.get(i).wrapFuture(each), eventFactory,
-                            futureCancelEventHandlerFactory, futureResultEventHandlerFactory);
+                            futureCancelWrapper, futureResultWrapper);
                     futWrappers.add(instrumentedFuture);
                     i++;
                 }
@@ -258,7 +257,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
         for (Callable<T> each: tasks) {
             final Callable<T> decoratedTask = typedCallableDecorator.wrapCallable(each);
             final CallableWrapper<T, ExecutionEvent> eachInstrumentedTask = new CallableWrapper<>(
-                    decoratedTask, eventFactory, callableExecutionEventHandlerFactory);
+                    decoratedTask, eventFactory, callableExecutionWrapper);
             decoratedTasks.add(decoratedTask);
             instrumentedTasks.add(eachInstrumentedTask);
         }
@@ -284,7 +283,7 @@ public class ExecutorServiceWrapper<ThreadPoolEvent, FutureEvent, ExecutionEvent
         for (Callable<T> each: tasks) {
             final Callable<T> decoratedTask = typedCallableDecorator.wrapCallable(each);
             final CallableWrapper<T, ExecutionEvent> eachInstrumentedTask = new CallableWrapper<>(
-                    decoratedTask, eventFactory, callableExecutionEventHandlerFactory);
+                    decoratedTask, eventFactory, callableExecutionWrapper);
             decoratedTasks.add(decoratedTask);
             instrumentedTasks.add(eachInstrumentedTask);
         }