+++ src/preflex/instrument.clj
@@ -11,6 +11,7 @@
   (:require
     [preflex.invokable :as invokable]
     [preflex.internal  :as in]
+    [preflex.task      :as task]
     [preflex.util      :as u])
   (:import
     [java.util UUID]
@@ -20,10 +21,13 @@
      CallableDecorator
      ConcurrentEventFactory
      ConcurrentEventHandlerFactory
+     ConcurrentTaskWrapper
      ExecutorServiceWrapper
      RunnableDecorator
      SharedContextCallable
-     SharedContextRunnable]))
+     SharedContextRunnable]
+    [preflex.instrument.task
+     Wrapper]))
 
 
 (defn make-event-handler
@@ -174,8 +178,8 @@
     (wrapRunnable [this runnable] (SharedContextRunnable. runnable (atom {})))))
 
 
-(defn make-shared-context-thread-pool-event-handlers
-  "Given keyword arguments in a map, return thread-pool instrumentation event handlers."
+(defn make-shared-context-thread-pool-task-wrappers
+  "Given keyword arguments in a map, return thread-pool instrumentation task wrapper fns."
   [{:keys [now-fn
            k-submit-begin
            k-submit-end
@@ -231,50 +235,50 @@
                              (swap! context-atom
                                (fn [context]
                                  (assoc context context-k (now-fn)))))))]
-    {:on-callable-submit  {:before (partial assoc-context  :callable k-submit-begin)
-                           :after  (partial after-submit   :callable)}
-     :on-runnable-submit  {:before (partial assoc-context  :runnable k-submit-begin)
-                           :after  (partial after-submit   :runnable)}
-     :on-callable-execute {:before (partial before-execute :callable)
-                           :after  (partial after-execute  :callable)}
-     :on-runnable-execute {:before (partial before-execute :runnable)
-                           :after  (partial after-execute  :runnable)}
-     :on-future-cancel    {:before (partial assoc-context  :future   k-future-cancel-begin)
-                           :after  (partial assoc-context  :future   k-future-cancel-end)}
-     :on-future-result    {:before (partial assoc-context  :future   k-future-result-begin)
-                           :after  (partial assoc-context  :future   k-future-result-end)}}))
+    {:callable-submit-wrapper  (fn [event f] (try       (assoc-context  :callable k-submit-begin event)        (f)
+                                               (finally (after-submit   :callable event))))
+     :runnable-submit-wrapper  (fn [event f] (try       (assoc-context  :runnable k-submit-begin event)        (f)
+                                       (finally (after-submit   :runnable event))))
+     :callable-execute-wrapper (fn [event f] (try       (before-execute :callable event)                       (f)
+                                       (finally (after-execute  :callable event))))
+     :runnable-execute-wrapper (fn [event f] (try       (before-execute :runnable event)                       (f)
+                                       (finally (after-execute  :runnable event))))
+     :future-cancel-wrapper    (fn [event f] (try       (assoc-context  :future   k-future-cancel-begin event) (f)
+                                       (finally (assoc-context  :future   k-future-cancel-end   event))))
+     :future-result-wrapper    (fn [event f] (try       (assoc-context  :future   k-future-result-begin event) (f)
+                                       (finally (assoc-context  :future   k-future-result-end   event))))}))
 
 
-(def shared-context-thread-pool-event-handlers-nanos  (make-shared-context-thread-pool-event-handlers
-                                                        {:now-fn                u/now-nanos
-                                                         :k-submit-begin        :submit-begin-ns
-                                                         :k-submit-end          :submit-end-ns
-                                                         :k-duration-submit     :duration-submit-ns
-                                                         :k-execute-begin       :execute-begin-ns
-                                                         :k-execute-end         :execute-end-ns
-                                                         :k-duration-queue      :duration-queue-ns
-                                                         :k-duration-execute    :duration-execute-ns
-                                                         :k-duration-response   :duration-response-ns
-                                                         :k-future-cancel-begin :cancel-begin-ns
-                                                         :k-future-cancel-end   :cancel-end-ns
-                                                         :k-future-result-begin :result-begin-ns
-                                                         :k-future-result-end   :result-end-ns}))
+(def shared-context-thread-pool-task-wrappers-nanos  (make-shared-context-thread-pool-task-wrappers
+                                                       {:now-fn                u/now-nanos
+                                                        :k-submit-begin        :submit-begin-ns
+                                                        :k-submit-end          :submit-end-ns
+                                                        :k-duration-submit     :duration-submit-ns
+                                                        :k-execute-begin       :execute-begin-ns
+                                                        :k-execute-end         :execute-end-ns
+                                                        :k-duration-queue      :duration-queue-ns
+                                                        :k-duration-execute    :duration-execute-ns
+                                                        :k-duration-response   :duration-response-ns
+                                                        :k-future-cancel-begin :cancel-begin-ns
+                                                        :k-future-cancel-end   :cancel-end-ns
+                                                        :k-future-result-begin :result-begin-ns
+                                                        :k-future-result-end   :result-end-ns}))
 
 
-(def shared-context-thread-pool-event-handlers-millis (make-shared-context-thread-pool-event-handlers
-                                                        {:now-fn                u/now-millis
-                                                         :k-submit-begin        :submit-begin-ms
-                                                         :k-submit-end          :submit-end-ms
-                                                         :k-duration-submit     :duration-submit-ms
-                                                         :k-execute-begin       :execute-begin-ms
-                                                         :k-execute-end         :execute-end-ms
-                                                         :k-duration-queue      :duration-queue-ms
-                                                         :k-duration-execute    :duration-execute-ms
-                                                         :k-duration-response   :duration-response-ms
-                                                         :k-future-cancel-begin :cancel-begin-ms
-                                                         :k-future-cancel-end   :cancel-end-ms
-                                                         :k-future-result-begin :result-begin-ms
-                                                         :k-future-result-end   :result-end-ms}))
+(def shared-context-thread-pool-task-wrappers-millis (make-shared-context-thread-pool-task-wrappers
+                                                       {:now-fn                u/now-millis
+                                                        :k-submit-begin        :submit-begin-ms
+                                                        :k-submit-end          :submit-end-ms
+                                                        :k-duration-submit     :duration-submit-ms
+                                                        :k-execute-begin       :execute-begin-ms
+                                                        :k-execute-end         :execute-end-ms
+                                                        :k-duration-queue      :duration-queue-ms
+                                                        :k-duration-execute    :duration-execute-ms
+                                                        :k-duration-response   :duration-response-ms
+                                                        :k-future-cancel-begin :cancel-begin-ms
+                                                        :k-future-cancel-end   :cancel-end-ms
+                                                        :k-future-result-begin :result-begin-ms
+                                                        :k-future-result-end   :result-end-ms}))
 
 
 (defn instrument-thread-pool
@@ -284,15 +288,15 @@
   ;; event generator
   :event-generator    instance of preflex.instrument.concurrent.ConcurrentEventFactory
 
-  ;; event handlers
-  on-callable-submit  instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-multiple-submit  instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-runnable-submit  instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-shutdown-request instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-callable-execute instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-runnable-execute instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-future-cancel    instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
-  on-future-result    instance of preflex.instrument.EventHandlerFactory or `event-handler-opts->factory` arg
+  ;; task wrappers
+  :callable-submit-wrapper  instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :multiple-submit-wrapper  instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :runnable-submit-wrapper  instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :shutdown-request-wrapper instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :callable-execute-wrapper instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :runnable-execute-wrapper instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :future-cancel-wrapper    instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
+  :future-result-wrapper    instance of preflex.instrument.task.Wrapper or argument to `preflex.task/make-wrapper`
 
   ;; decorators
   :callable-decorator instance of preflex.instrument.concurrent.CallableDecorator
@@ -307,43 +311,44 @@
                                         runnable-decorator
                                         ;; event generator
                                         event-generator
-                                        ;; event handlers
-                                        on-callable-submit
-                                        on-multiple-submit
-                                        on-runnable-submit
-                                        on-shutdown-request
-                                        on-callable-execute
-                                        on-runnable-execute
-                                        on-future-cancel
-                                        on-future-result]
+                                        ;; task wrappers
+                                        callable-submit-wrapper
+                                        multiple-submit-wrapper
+                                        runnable-submit-wrapper
+                                        shutdown-request-wrapper
+                                        callable-execute-wrapper
+                                        runnable-execute-wrapper
+                                        future-cancel-wrapper
+                                        future-result-wrapper]
                                  :or {;; decorators
                                       callable-decorator CallableDecorator/IDENTITY
                                       runnable-decorator RunnableDecorator/IDENTITY
                                       ;; event generator
                                       event-generator    default-thread-pool-event-generator
-                                      ;; event handlers
-                                      on-callable-submit  EventHandlerFactory/NOP
-                                      on-multiple-submit  EventHandlerFactory/NOP
-                                      on-runnable-submit  EventHandlerFactory/NOP
-                                      on-shutdown-request EventHandlerFactory/NOP
-                                      on-callable-execute EventHandlerFactory/NOP
-                                      on-runnable-execute EventHandlerFactory/NOP
-                                      on-future-cancel    EventHandlerFactory/NOP
-                                      on-future-result    EventHandlerFactory/NOP}
+                                      ;; task wrappers
+                                      callable-submit-wrapper  Wrapper/IDENTITY
+                                      multiple-submit-wrapper  Wrapper/IDENTITY
+                                      runnable-submit-wrapper  Wrapper/IDENTITY
+                                      shutdown-request-wrapper Wrapper/IDENTITY
+                                      callable-execute-wrapper Wrapper/IDENTITY
+                                      runnable-execute-wrapper Wrapper/IDENTITY
+                                      future-cancel-wrapper    Wrapper/IDENTITY
+                                      future-result-wrapper    Wrapper/IDENTITY
+                                      }
                                  :as opts}]
   (ExecutorServiceWrapper.
     thread-pool
     event-generator
-    (let [as-event-handler-factory #(if (instance? EventHandlerFactory %) % (event-handler-opts->factory %))]
-      (ConcurrentEventHandlerFactory.
-        (as-event-handler-factory on-callable-submit)
-        (as-event-handler-factory on-multiple-submit)
-        (as-event-handler-factory on-runnable-submit)
-        (as-event-handler-factory on-shutdown-request)
-        (as-event-handler-factory on-callable-execute)
-        (as-event-handler-factory on-runnable-execute)
-        (as-event-handler-factory on-future-cancel)
-        (as-event-handler-factory on-future-result)))
+    (let [as-task-wrapper (fn ^Wrapper [x] (if (fn? x) (task/make-wrapper x) x))]
+      (ConcurrentTaskWrapper.
+        (as-task-wrapper callable-submit-wrapper)
+        (as-task-wrapper multiple-submit-wrapper)
+        (as-task-wrapper runnable-submit-wrapper)
+        (as-task-wrapper shutdown-request-wrapper)
+        (as-task-wrapper callable-execute-wrapper)
+        (as-task-wrapper runnable-execute-wrapper)
+        (as-task-wrapper future-cancel-wrapper)
+        (as-task-wrapper future-result-wrapper)))
     ;; decorators
     callable-decorator
     runnable-decorator))