+++ test/preflex/instrument_test.clj
@@ -40,7 +40,7 @@
   (testing "shared context instrumentation"
     (with-active-thread-pool [^ExecutorService thread-pool (core/make-bounded-thread-pool 10 10)]
       (let [instru-pool (instru/instrument-thread-pool thread-pool
-                          (-> instru/shared-context-thread-pool-event-handlers-millis
+                          (-> instru/shared-context-thread-pool-task-wrappers-millis
                             (assoc
                               :callable-decorator  instru/default-shared-context-callable-decorator
                               :runnable-decorator  instru/default-shared-context-runnable-decorator)))]
@@ -78,7 +78,7 @@
     (with-active-thread-pool [^ExecutorService thread-pool (core/make-bounded-thread-pool 10 10)]
       (let [invoker (fn [g context-atom] (swap! context-atom assoc :added-by-invoker 20) (g))
             instru-pool (instru/instrument-thread-pool thread-pool
-                          (-> instru/shared-context-thread-pool-event-handlers-nanos
+                          (-> instru/shared-context-thread-pool-task-wrappers-nanos
                             (assoc
                               :callable-decorator  (instru/make-shared-context-callable-decorator invoker)
                               :runnable-decorator  (instru/make-shared-context-runnable-decorator invoker))))]